requirements.txt
discord.py
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
PyYAML
google-generativeai